+++ CHANGES.md
@@ -3,6 +3,7 @@
 #### next release (8.3.5)
 
 - Add `includeMembersRegex` to `GroupTraits`. This can be used to filter group members by id/name using a regular expression.
+- `BingMapsCatalogItem` now supports Bing's `culture` parameter.
 - [The next improvement]
 
 #### 8.3.4 - 2023-09-15